blemesh-rssi/collector/requirements.txt
pyserial
cbor2
ujson